Fix: 이미지 안보이는 에러 수정, StrictMode로 인한 useEffect 두 번 호출 이슈
