+++ src/services/api/pages/pageContent.ts
@@ -2,7 +2,7 @@ import { AxiosResponse } from 'axios';
 import Instance from '../Instance';
 import PATH from '@/constants/path/Path';
 
-export default async function pageContent(path: string): Promise<object> {
+export default async function pageContent(path: string) {
   const response: AxiosResponse = await Instance.get(`${PATH.API.PAGES.page}/${path}`);
   const result = {
     content: '',
@@ -14,10 +14,9 @@ export default async function pageContent(path: string): Promise<object> {
     const { content } = response.data.data.editPage;
     const domainString: string = response.data.data.pageUrl;
     const url = new URL(domainString);
-    const domain = url.hostname.split('.')[0];
 
     result.content = content;
-    result.domain = domain;
+    result.domain = url.hostname;
   }
 
   return result;
